Refactor dexie.js - Separated buklod-specific code to its app-level version - Create a new collection to accomodate local imports
