+++ js/dexie.js
@@ -35,7 +35,7 @@ const collections = {
 }
 
 
-const buklodSchema = {
+export const buklodSchema = {
   version: 0,
   type: 'object',
   primaryKey: 'id',
@@ -103,6 +103,7 @@ export async function initDb(uid) {
 
   await dbPromise.addCollections({
     buklod: { schema: buklodSchema },
+    buklodImport: {schema: buklodSchema}, // Collection exclusively used for data imports
     evacCenters: { schema: evacCentersSchema }
   });
 
@@ -116,7 +117,7 @@ export function startFirestoreSync(db, uid) {
   console.log(`Syncing ${collections["buklod-tao"].households} with firestore...`)
   db.buklodSyncState = replicateRxCollection({ 
     collection: db.buklod,
-    replicationIdentifier: 'buklod-test-sync-v4',
+    replicationIdentifier: 'buklod-test-sync-v9',
     live: true, 
     retryTime: 5 * 1000, 
     
@@ -265,117 +266,3 @@ export function startFirestoreSync(db, uid) {
     console.error("Evac Centers Sync Error:", err);
   });
 }
-
-export async function resetDatabase() {
-  if (!dbPromise) return;
-  const db = await dbPromise;
-  await cancelReplication(db);
-  await db.remove();
-  dbPromise = null;
-  dbUid = null;
-} 
-
-export async function cancelReplication(db) {
-  await Promise.allSettled([
-    db.evacSyncState?.cancel(),
-    db.buklodSyncState?.cancel(),
-  ]);
-  if (!IS_TESTING) console.log("Replication engine has been cancelled.");
-}
-
-export async function parseData(file) {
-  try {
-    const parseRow = (row) => {
-      const get = (header, default_val = "") => {
-        return row[header] == null ? default_val : String(row[header]).trim();
-      }
-      const num = (header) => {
-        const raw = get(header).replace(/,/g,'');
-        const n = Number(raw);
-        return Number.isFinite(n) ? n : 0;
-      }
-      const coord = () => {
-        const lat = num("Latitude");
-        const lng = num("Longitude");
-        if (lat === 0 && lng === 0) { return null }
-        else { return {_lat: lat, _lng: lng}; }
-      }
-
-      return {
-        id: String(generateHash(get("Household Name"))),
-        household_name: get("Household Name"),
-        household_phase: get("Phase"),
-        street: get("Street"),
-        household_address: get("Address"),
-        contact_number: get("Contact Number"),
-        residency_status: get("Residency Status"),
-        is_hoa_noa: get("HOA/NOA/Others"),
-        household_material: get("House Material"),
-        nearest_evac: get("Nearest Evacuation Center"),
-        location_link: get("Location Link"),
-        location_coordinates: coord(),
-        
-        number_residents: num("Number of Residents"),
-        number_minors: num("Minors"),
-        number_seniors: num("Seniors"),
-        number_pwd: num("PWD"),
-        number_sick: num("Sick"),
-        number_pregnant: num("Pregnant"),
-        sickness_present: get("Sickness Present"),
-        number_families: num("Number of Families"),
-        number_healthy: num("Healthy"),
-        
-        earthquake_risk: get("Earthquake Risk"),
-        earthquake_risk_description: get("Earthquake Description"),
-        fire_risk: get("Fire Risk"),
-        fire_risk_description: get("Fire Description"),
-        flood_risk: get("Flood Risk"),
-        flood_risk_description: get("Flood Description"),
-        landslide_risk: get("Landslide Risk"),
-        landslide_risk_description: get("Landslide Description"),
-        storm_risk: get("Storm Risk"),
-        storm_risk_description: get("Storm Description"),
-        
-        exit_points: num("Exit Points"),
-        disaster_response_plan: get("Disaster Response Plan"),
-        knowledge_readiness: num("Knowledge Readiness"),
-        before_disaster_actions: get("Before Disaster Actions"),
-        during_disaster_actions: get("During Disaster Actions"),
-        after_disaster_actions: get("After Disaster Actions"),
-        important_notes: get("Important Notes"),
-        notes: get("Notes"),
-        
-        source_dataset: get("Source Dataset"),
-        
-        updatedAt: Date.now(),
-        _deleted: false
-      }
-    }
-    const data = await file.arrayBuffer();
-    const workbook = XLSX.read(data);
-    const masterSheet = workbook.Sheets["Master Sheet"];
-    if (!masterSheet) {
-      throw new Error("Spreadsheet is missing a 'Master Sheet' tab.");
-    }
-    const jsonData = XLSX.utils.sheet_to_json(masterSheet);
-
-    return jsonData
-      .filter((r) => String(r["Household Name"] ?? "").trim() !== "")
-      .map(parseRow);
-  } catch (err) {
-    console.error("Import failed:", err);
-    throw new Error(`Could not parse file: ${err.message}`);
-  }
-}
-
-// Source - https://stackoverflow.com/a/7616484
-// Posted by esmiralha, modified by community. See post 'Timeline' for change history
-// Retrieved 2026-06-17, License - CC BY-SA 4.0
-function generateHash(string) {
-  let hash = 0;
-  for (const char of string) {
-    hash = (hash << 5) - hash + char.charCodeAt(0);
-    hash |= 0; // Constrain to 32bit integer
-  }
-  return hash;
-};
